+++ ffw_bringup/launch/ffw_pack.launch.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 #
-# Copyright 2024 ROBOTIS CO., LTD.
+# Copyright 2026 ROBOTIS CO., LTD.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,34 +14,40 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-# Author: Sungho Woo, Woojin Wie
-
-import os
+# Author: Hyungyu Kim
 
 from ament_index_python.packages import get_package_share_directory
 from launch import LaunchDescription
+from launch.actions import DeclareLaunchArgument
 from launch.actions import IncludeLaunchDescription
 from launch.actions import LogInfo
 from launch.launch_description_sources import PythonLaunchDescriptionSource
+from launch.substitutions import LaunchConfiguration
+
+
+VALID_ROBOT_TYPES = ['sg2', 'bg2', 'sh5', 'bh5']
 
 
 def generate_launch_description():
-    # Get the package share directory
     pkg_share = get_package_share_directory('ffw_bringup')
 
-    # Path to the omy_3m.launch.py file
-    ffw_bg2_pack_launch_file = os.path.join(pkg_share, 'launch', 'ffw_bg2_follower_ai.launch.py')
+    robot_type = LaunchConfiguration('type')
 
-    # Include the ffw_bg2_follower_ai.launch.py with pack parameters
-    ffw_bg2_pack_launch = IncludeLaunchDescription(
-        PythonLaunchDescriptionSource([ffw_bg2_pack_launch_file]),
+    pack_launch = IncludeLaunchDescription(
+        PythonLaunchDescriptionSource([
+            pkg_share, '/launch/ffw_', robot_type, '_follower_ai.launch.py'
+        ]),
         launch_arguments={
             'init_position': 'true',
-            'init_position_file': 'pack_position.yaml'
-        }.items()
+            'init_position_file': 'pack_position.yaml',
+        }.items(),
     )
 
     return LaunchDescription([
-        LogInfo(msg='Starting FFW BG2 packing...'),
-        ffw_bg2_pack_launch,
+        DeclareLaunchArgument(
+            'type',
+            description=f'Robot type to launch pack position. One of: {VALID_ROBOT_TYPES}',
+        ),
+        LogInfo(msg=['Starting FFW ', robot_type, ' packing...']),
+        pack_launch,
     ])